+++ frontend/src/components/layouts/common/TextAlignLayout.tsx
